Excessive stock out number warning added
If a user tries to consume more than what is in stock, a warning is now
being issued using a modal to check the stock out number.
An additional navigation home added to the sign out button. It is now
navigating to the home screen instead of login page.
Styling added to the stock out button to make it wider on smaller
screens.